requirements.txt
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
decorator==4.4.2
et-xmlfile==1.1.0
future==0.18.2
h11==0.13.0
huepy==1.2.1
idna==3.3
imageio==2.15.0
imageio-ffmpeg==0.4.5
imgkit==1.2.2
iniconfig==1.1.1
instabot==0.117.0
instagram-basic-display==1.0.5
Jinja2==3.0.3
MarkupSafe==2.0.1
mock==4.0.3
moviepy==1.0.3
mysql-connector-python==8.0.28
numpy==1.22.2
openai==0.14.0
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.0
pandas-stubs==1.2.0.47
Pillow==9.0.1
pluggy==1.0.0
proglog==0.1.9
protobuf==3.19.4
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
responses==0.18.0
schedule==1.1.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.5.2
wsproto==1.0.0
